Refactor: Implement operations page changes
This commit implements the changes requested for the operations page.
